+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add Instagram post embeds
 - Add setting to enable Instagram embeds, off by default
 - Make embedded tweet webview background transparent
+- Upload progress indicator
 
 ## v1.0.46